Moved NewAlertStore and NewLegacyAlertStore from config.go to store.go
